# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a todo management API for couples. It allows two users (anyel and alexis) to create, manage, and complete todos together.",
        "title": "LoveApp Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@loveapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change the password for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "Password change request",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ChangePasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate user with username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is running",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of todos with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "List todos",
                "parameters": [
                    {
                        "enum": [
                            "all",
                            "pending",
                            "completed"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by creator user ID",
                        "name": "creator_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todos retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TodoListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new todo item (only for authorized users: anyel, alexis)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "parameters": [
                    {
                        "description": "Todo creation request",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTodoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Todo created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTodoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a todo (only the creator can delete their todos)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/todos/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a todo as completed by the authenticated user (both users must mark it to be fully completed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Mark todo as completed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo marked as completed",
                        "schema": {
                            "$ref": "#/definitions/models.CompleteTodoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ChangePasswordRequest": {
            "description": "Change password request",
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "New password",
                    "type": "string",
                    "example": "newpassword"
                },
                "old_password": {
                    "description": "Current password",
                    "type": "string",
                    "example": "oldpassword"
                }
            }
        },
        "models.ChangePasswordResponse": {
            "description": "Change password response",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Contraseña cambiada exitosamente"
                }
            }
        },
        "models.CompleteTodoResponse": {
            "description": "Complete todo response",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Todo marcado como completado"
                },
                "todo": {
                    "description": "Updated todo",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Todo"
                        }
                    ]
                }
            }
        },
        "models.CreateTodoRequest": {
            "description": "Create todo request",
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "description": "Todo description",
                    "type": "string",
                    "example": "Ir al supermercado y comprar frutas y verduras"
                },
                "title": {
                    "description": "Todo title",
                    "type": "string",
                    "example": "Comprar comida"
                }
            }
        },
        "models.CreateTodoResponse": {
            "description": "Create todo response",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Todo creado exitosamente"
                },
                "todo": {
                    "description": "Created todo",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Todo"
                        }
                    ]
                }
            }
        },
        "models.ErrorResponse": {
            "description": "Error response",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Error message"
                },
                "message": {
                    "description": "Detailed error message",
                    "type": "string",
                    "example": "Detailed error"
                }
            }
        },
        "models.HealthResponse": {
            "description": "Health check response",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Status message",
                    "type": "string",
                    "example": "Servicio funcionando"
                },
                "status": {
                    "description": "Service status",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.LoginRequest": {
            "description": "Login credentials",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password",
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "anyel"
                }
            }
        },
        "models.LoginResponse": {
            "description": "Login response with token and user info",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Inicio de sesión exitoso"
                },
                "token": {
                    "description": "JWT token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "description": "User information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
        "models.SuccessResponse": {
            "description": "Success response",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Optional data"
                },
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Operación exitosa"
                }
            }
        },
        "models.Todo": {
            "description": "Todo item information",
            "type": "object",
            "properties": {
                "completed_alexis": {
                    "description": "Completed by Alexis",
                    "type": "boolean",
                    "example": false
                },
                "completed_anyel": {
                    "description": "Completed by Anyel",
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "creator_id": {
                    "description": "Creator user ID",
                    "type": "integer",
                    "example": 1
                },
                "creator_username": {
                    "description": "Creator username",
                    "type": "string",
                    "example": "anyel"
                },
                "description": {
                    "description": "Todo description",
                    "type": "string",
                    "example": "Ir al supermercado y comprar..."
                },
                "id": {
                    "description": "Todo ID",
                    "type": "integer",
                    "example": 1
                },
                "is_completed": {
                    "description": "Overall completion status",
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "description": "Todo title",
                    "type": "string",
                    "example": "Comprar comida"
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "models.TodoListResponse": {
            "description": "Todo list response",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Todos obtenidos exitosamente"
                },
                "todos": {
                    "description": "List of todos",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Todo"
                    }
                },
                "total": {
                    "description": "Total count",
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.User": {
            "description": "User account information",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "description": "User ID",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "anyel"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}